Backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# Celery
celery==5.3.6
redis==5.0.1

# LLM Providers
openai==1.10.0
anthropic==0.8.1
mistralai==1.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
ipython==8.20.0
django-extensions==3.2.3

# WSGI Server
gunicorn==21.2.0

# Monitoring
psutil==5.9.8
